backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
sqlalchemy[asyncio]
alembic
asyncpg
psycopg2-binary
supabase
python-jose[cryptography]
passlib[bcrypt]
